Testing Devise member registrations
